frontend/py_scripts/requirements.txt
# vlabs/public/python-scripts/requirements.txt
junos-eznc
paramiko
PyYAML
jsnapy==1.3.8
lxml
tabulate
psutil
colorama
jinja2
cryptography
flask
werkzeug
fastapi
uvicorn
aiohttp
requests
